CAP-2684 Add unit tests for Metadata and Manifest chunking and buffering
